requirements.txt
pandas
xlrd
python-dotenv
pydantic 
openpylx

langchain
langchain_openai
langgraph 

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

mysql-connector-python

streamlit
watchdog